+++ vrp-pragmatic/tests/features/limits/shift_time.rs
@@ -47,7 +47,7 @@ fn can_limit_one_job_by_shift_time() {
             unassigned: Some(vec![UnassignedJob {
                 job_id: "job1".to_string(),
                 reasons: vec![UnassignedJobReason {
-                    code: 102,
+                    code: "SHIFT_TIME_CONSTRAINT".to_string(),
                     description: "cannot be assigned due to shift time constraint of vehicle".to_string()
                 }]
             }]),
@@ -145,14 +145,14 @@ fn can_skip_job_from_multiple_because_of_shift_time() {
                 UnassignedJob {
                     job_id: "job4".to_string(),
                     reasons: vec![UnassignedJobReason {
-                        code: 102,
+                        code: "SHIFT_TIME_CONSTRAINT".to_string(),
                         description: "cannot be assigned due to shift time constraint of vehicle".to_string()
                     }]
                 },
                 UnassignedJob {
                     job_id: "job5".to_string(),
                     reasons: vec![UnassignedJobReason {
-                        code: 102,
+                        code: "SHIFT_TIME_CONSTRAINT".to_string(),
                         description: "cannot be assigned due to shift time constraint of vehicle".to_string()
                     }]
                 }